CLDIDE-147 Merge testsuite IDE-1.2.0-M5-branch with ide-1.2.0-m5-trunk

+++ exo-ide-testsuite-selenium/src/test/java/org/exoplatform/ide/operation/restservice/RESTServiceGetURLTest.java
@@ -81,7 +81,7 @@ public class RESTServiceGetURLTest extends BaseTest
       IDE.REST_SERVICE.openGetURLForm();
 
       String url = IDE.REST_SERVICE.getUrlFromGetURLForm();
-
+      
       assertTrue((BASE_URL + REST_CONTEXT_IDE + "/testService").equals(url));
       //Close form
       IDE.REST_SERVICE.closeGetURLForm();
@@ -117,10 +117,13 @@ public class RESTServiceGetURLTest extends BaseTest
    {
       try
       {
-         //TODO     Utils.undeployService(BASE_URL, REST_CONTEXT, url);
+        //TODO     Utils.undeployService(BASE_URL, REST_CONTEXT, url);
+        if (IDE.OUTPUT.getOutputMessage(2).contains("/" + PROJECT + "/" + FILE_NAME + " undeployed successfully."))
+        {
          VirtualFileSystemUtils.delete(WS_URL + PROJECT);
+        }        
       }
-      catch (IOException e)
+        catch (IOException e)
       {
       }
    }
